refactor(ops-flow): update import paths for task-related schemas
