+++ apps/web/src/services/ops-flow/composables/use-task-type-field.ts
@@ -4,7 +4,6 @@ import { ref, computed } from 'vue';
 import { getTextHighlightRegex } from '@cloudforet/mirinae';
 import type { AutocompleteHandler, SelectDropdownMenuItem } from '@cloudforet/mirinae/types/controls/dropdown/select-dropdown/type';
 
-import type { TaskTypeModel } from '@/schema/opsflow/task-type/model';
 import { getParticle, i18n } from '@/translations';
 
 import { useFieldValidator } from '@/common/composables/form-validator';
@@ -13,6 +12,8 @@ import { useTaskTypeStore } from '@/services/ops-flow/stores/task-type-store';
 import {
     useTaskManagementTemplateStore,
 } from '@/services/ops-flow/task-management-templates/stores/use-task-management-template-store';
+
+import type { TaskTypeModel } from '@/api-clients/opsflow/task/schema-type/model';
 
 export const useTaskTypeField = ({
     isRequired, categoryId,
