+++ apps/web/src/services/ops-flow/stores/admin/task-category-page-store.ts
@@ -6,16 +6,17 @@ import {
 
 import { defineStore } from 'pinia';
 
-import type { TaskCategoryModel } from '@/schema/opsflow/task-category/model';
-import type { TaskTypeModel } from '@/schema/opsflow/task-type/model';
-import type { TaskModel } from '@/schema/opsflow/task/model';
-import type { TaskStatusOption, TaskStatusOptions, TaskStatusType } from '@/schema/opsflow/task/type';
+import type { TaskModel } from '@/api-clients/opsflow/task/schema/model';
+import type { TaskStatusOption, TaskStatusOptions, TaskStatusType } from '@/api-clients/opsflow/task/schema/type';
 
 import ErrorHandler from '@/common/composables/error/errorHandler';
 
 import { useTaskAPI } from '@/services/ops-flow/composables/use-task-api';
 import { useTaskCategoryStore } from '@/services/ops-flow/stores/task-category-store';
 import { useTaskTypeStore } from '@/services/ops-flow/stores/task-type-store';
+
+import type { TaskCategoryModel } from '@/api-clients/opsflow/task/schema-category/model';
+import type { TaskTypeModel } from '@/api-clients/opsflow/task/schema-type/model';
 
 interface UseTaskCategoryPageStoreState {
     currentCategoryId?: string;
