+++ apps/web/src/services/ops-flow/stores/board-page-store.ts
@@ -5,9 +5,10 @@ import {
 
 import { defineStore } from 'pinia';
 
-import type { TaskCategoryModel } from '@/schema/opsflow/task-category/model';
-
 import { useTaskCategoryStore } from '@/services/ops-flow/stores/task-category-store';
+
+import type { TaskCategoryModel } from '@/api-clients/opsflow/task/schema-category/model';
+
 
 interface UseBoardPageStoreGetters {
     currentCategory: ComputedRef<TaskCategoryModel|undefined>;
