+++ apps/web/src/services/ops-flow/task-fields-configuration/options-generator-templates/DropdownOptionsGenerator.vue
@@ -10,7 +10,7 @@ import {
     PFieldTitle, PButton, PIconButton, PI,
 } from '@cloudforet/mirinae';
 
-import type { DropdownTaskFieldOptions, TaskFieldEnum } from '@/schema/opsflow/_types/task-field-type';
+import type { DropdownTaskFieldOptions, TaskFieldEnum } from '@/api-clients/opsflow/_types/task-field-type';
 import { i18n } from '@/translations';
 
 import getRandomId from '@/lib/random-id-generator';
