+++ apps/web/src/services/ops-flow/task-fields-configuration/options-generator-templates/ParagraphOptionsGenerator.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import type { ParagraphTaskFieldOptions } from '@/schema/opsflow/_types/task-field-type';
+import type { ParagraphTaskFieldOptions } from '@/api-clients/opsflow/_types/task-field-type';
 
 import TextEditor from '@/common/components/editor/TextEditor.vue';
