+++ apps/web/src/services/ops-flow/task-fields-configuration/types/mutable-task-field-type.ts
@@ -1,4 +1,4 @@
-import type { TaskField } from '@/schema/opsflow/_types/task-field-type';
+import type { TaskField } from '@/api-clients/opsflow/_types/task-field-type';
 
 export type MutableTaskField = TaskField & {
     _field_id: string; // original field_id. immutable. it must be removed when saving
