+++ apps/web/src/services/ops-flow/task-fields-configuration/types/options-generator-type.ts
@@ -1,4 +1,4 @@
-import type { TaskFieldOptions } from '@/schema/opsflow/_types/task-field-type';
+import type { TaskFieldOptions } from '@/api-clients/opsflow/_types/task-field-type';
 
 export interface OptionsGeneratorProps<Options extends TaskFieldOptions> {
     options: Options;
