+++ apps/web/src/services/ops-flow/task-fields-configuration/types/task-field-dropdown-enum-type.ts
@@ -1,4 +1,4 @@
-import type { TaskFieldEnum } from '@/schema/opsflow/_types/task-field-type';
+import type { TaskFieldEnum } from '@/api-clients/opsflow/_types/task-field-type';
 
 export interface ControllableTaskFieldEnum extends TaskFieldEnum {
     _id: string;
