+++ apps/web/src/services/ops-flow/task-fields-configuration/types/task-field-type-metadata-type.ts
@@ -1,4 +1,4 @@
-import type { TaskFieldType } from '@/schema/opsflow/_types/task-field-type';
+import type { TaskFieldType } from '@/api-clients/opsflow/_types/task-field-type';
 
 import type { DEFAULT_FIELD_ID_MAP } from '@/services/ops-flow/task-fields-configuration/constants/default-field-constant';
