+++ apps/web/src/services/ops-flow/task-fields-form/field-templates/DropdownTaskField.vue
@@ -6,7 +6,7 @@ import {
 } from '@cloudforet/mirinae';
 import type { AutocompleteHandler, SelectDropdownMenuItem } from '@cloudforet/mirinae/types/controls/dropdown/select-dropdown/type';
 
-import type { DropdownTaskField } from '@/schema/opsflow/_types/task-field-type';
+import type { DropdownTaskField } from '@/api-clients/opsflow/_types/task-field-type';
 
 import { useTaskFieldValidation } from '@/services/ops-flow/task-fields-form/composables/use-task-field-validation';
 import type {
