+++ apps/web/src/services/ops-flow/task-fields-form/types/task-field-form-type.ts
@@ -1,5 +1,5 @@
+import type { TaskField } from '@/api-clients/opsflow/_types/task-field-type';
 import type { FileModel } from '@/schema/file-manager/model';
-import type { TaskField } from '@/schema/opsflow/_types/task-field-type';
 
 export interface References {
   project_id?: string;
